abstract out the utitity functions in another class
